// frontend/src/api/generated/model/aiPromptFlowResponseDtoFlowsItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */
import type { AiPromptFlowResponseDtoFlowsItemEdgesItem } from './aiPromptFlowResponseDtoFlowsItemEdgesItem';
import type { AiPromptFlowResponseDtoFlowsItemId } from './aiPromptFlowResponseDtoFlowsItemId';
import type { AiPromptFlowResponseDtoFlowsItemNodesItem } from './aiPromptFlowResponseDtoFlowsItemNodesItem';
import type { AiPromptFlowResponseDtoFlowsItemStagesItem } from './aiPromptFlowResponseDtoFlowsItemStagesItem';

export type AiPromptFlowResponseDtoFlowsItem = {
  id: AiPromptFlowResponseDtoFlowsItemId;
  name: string;
  description: string;
  stages: AiPromptFlowResponseDtoFlowsItemStagesItem[];
  nodes: AiPromptFlowResponseDtoFlowsItemNodesItem[];
  edges: AiPromptFlowResponseDtoFlowsItemEdgesItem[];
};